@@ -18,18 +18,18 @@
   },
   "require": {
     "php": "^8.0",
-    "psr/log": "^1.0",
-    "php-http/discovery": "^1.13",
+    "psr/log": "^3.0.1",
+    "php-http/discovery": "^1.19.0",
     "psr/http-client-implementation": "^1.0",
     "psr/http-factory-implementation": "^1.0",
-    "php-http/client-common": "^2.3",
-    "knplabs/github-api": "^3.3"
+    "php-http/client-common": "^2.7.1",
+    "knplabs/github-api": "^3.16.0"
   },
   "require-dev": {
-    "symfony/dom-crawler": "^5.3",
-    "symfony/css-selector": "^5.3",
-    "guzzlehttp/guzzle": "^7.1",
-    "guzzlehttp/psr7": "2.0.0-rc1"
+    "symfony/dom-crawler": "^7.1.1",
+    "symfony/css-selector": "^7.1.1",
+    "guzzlehttp/guzzle": "^7.9.2",
+    "guzzlehttp/psr7": "2.7.0"
   },
   "suggest": {
     "symfony/dom-crawler": "Used by Github Inspector class",
@@ -38,6 +38,11 @@
   "autoload": {
     "psr-4": {
       "Github\\Utils\\": "src/"
+    }
+  },
+  "config": {
+    "allow-plugins": {
+      "php-http/discovery": true
     }
   }
 }
