@@ -14,17 +14,17 @@ interface GithubTokenFactoryInterface
     /**
      * Creates a token or list of tokens.
      *
-     * @param array|string $type   Token type or an array of token definitions
+     * @param array|string $type Token type or an array of token definitions
      * @param ...          $params Token params
      *
      * @return GithubTokenInterface|GithubTokenInterface[]
      */
-    public static function create(string | array $type, ...$params): array | GithubTokenInterface;
+    public static function create(array|string $type, ...$params): array|GithubTokenInterface;
 
     /**
      * Gets a list of supported token types or checks whether the given type is supported.
      *
      * @param null|string $type Token type to check against
      */
-    public static function supports(string $type = null): bool | array;
+    public static function supports(?string $type = null): array|bool;
 }
